Add UriFactory::createFromSapi()
This method will provide a public API to create Uri instances from
SERVER context. This helps libraries that depend on diactoros and have
their own ServerRequest or Uri implementations use the logic in
diactoros more efficiently.
